requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==4.1.7
django-cors-headers==3.14.0
django-simple-history==3.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.3.0
drf-spectacular==0.26.1
inflection==0.5.1
jsonschema==4.17.3
PyJWT==1.7.1
pymongo==4.3.3
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.2.4
tzdata==2022.7
uritemplate==4.1.1
